+++ client/django-formset/DualSelector.ts
@@ -21,7 +21,7 @@ export class DualSelector extends IncompleteSelect {
 	private readonly redoButton?: HTMLButtonElement;
 	private historicalValues: string[][] = [];
 	private historyCursor: number = 0;
-	private lastRemoteQuery?: string;
+	private lastRemoteQuery = new URLSearchParams();
 	private readonly renderNoResults: Function;
 
 	constructor(selectorElement: HTMLSelectElement, name: string) {
@@ -65,6 +65,7 @@ export class DualSelector extends IncompleteSelect {
 		this.initialize();
 		this.transferStyles();
 		this.setButtonsState();
+		this.setupFilters(selectorElement);
 	}
 
 	private installEventHandlers() {
@@ -84,45 +85,30 @@ export class DualSelector extends IncompleteSelect {
 
 	private initialize() {
 		const initialValues: string[] = [];
-		const optionElemens = this.selectorElement.querySelectorAll(':scope > option') as NodeListOf<HTMLOptionElement>;
-		optionElemens.forEach(option => {
-			const clone = option.cloneNode() as HTMLOptionElement;
+		const optionElements = this.selectorElement.querySelectorAll(':scope > option') as NodeListOf<HTMLOptionElement>;
+		optionElements.forEach(option => {
+			const optionData : OptionData = {id: option.value, label: option.label};
 			if (option.selected) {
-				option.selected = false;
-				this.selectRightElement.add(option);
+				this.addOptionToSelectElement(optionData, this.selectRightElement);
 				initialValues.push(option.value);
 			} else {
-				this.selectLeftElement.add(option);
+				this.addOptionToSelectElement(optionData, this.selectLeftElement);
 			}
-			this.selectorElement.add(clone);  // add clone without inner text
+			option.replaceChildren();  // selectorElement is shadowed
 		});
 		const optionGroupElements = this.selectorElement.querySelectorAll(':scope > optgroup') as NodeListOf<HTMLOptGroupElement>;
 		optionGroupElements.forEach(optGroupElement => {
-			const selectedOptions = optGroupElement.querySelectorAll('option:checked') as NodeListOf<HTMLOptionElement>;
-			if (selectedOptions.length > 0) {
-				const newOptGroup = document.createElement('optgroup');
-				newOptGroup.label = optGroupElement.label;
-				this.selectRightElement.add(newOptGroup);
-				selectedOptions.forEach(option => {
-					const clone = option.cloneNode() as HTMLOptionElement;
-					option.selected = false;
-					newOptGroup.appendChild(option);
+			optGroupElement.querySelectorAll('option').forEach(option => {
+				const optionData : OptionData = {id: option.value, label: option.label, optgroup: optGroupElement.label};
+				if (option.selected) {
+					this.addOptionToSelectElement(optionData, this.selectRightElement);
 					initialValues.push(option.value);
-					this.selectorElement.add(clone);  // add clone without inner text
-				});
-			}
-			const unselectedOptions = optGroupElement.querySelectorAll('option:not(:checked)');
-			if (unselectedOptions.length > 0) {
-				const newOptGroup = document.createElement('optgroup');
-				newOptGroup.label = optGroupElement.label;
-				this.selectLeftElement.add(newOptGroup);
-				unselectedOptions.forEach(option => {
-					const clone = option.cloneNode() as HTMLOptionElement;
-					newOptGroup.appendChild(option);
-					this.selectorElement.add(clone);  // add clone without inner text
-				});
-			}
-			this.selectorElement.removeChild(optGroupElement);
+				} else {
+					this.addOptionToSelectElement(optionData, this.selectLeftElement);
+				}
+				option.replaceChildren();  // selectorElement is shadowed
+			});
+			optGroupElement.replaceWith(...optGroupElement.childNodes);
 		});
 		this.historicalValues.push(initialValues);
 		this.setHistoryCursor(0);
@@ -137,50 +123,57 @@ export class DualSelector extends IncompleteSelect {
 		selectElement.add(option);
 	}
 
-	private addOptionToSelectElement(option: OptionData, target: HTMLSelectElement | SortableSelectElement) {
-		if (this.selectorElement.querySelector(`option[value="${option.id}"]`))
-			return;
+	private addOptionToSelectElement(option: OptionData, target: HTMLSelectElement | SortableSelectElement) : HTMLOptionElement {
+		const optionElement = target.querySelector(`option[value="${option.id}"]`);
+		if (optionElement instanceof HTMLOptionElement)
+			return optionElement;  // prevent duplicates
 		const newOptionElement = new Option(option.label, option.id);
 		if (typeof option.optgroup === 'string') {
-			const optGroupElement = target.querySelector(`optgroup[label="${option.optgroup}"]`);
-			if (optGroupElement instanceof HTMLOptGroupElement) {
-				optGroupElement.appendChild(newOptionElement);
-			} else {
-				const newOptGroupElement = document.createElement('optgroup');
-				newOptGroupElement.label = option.optgroup;
-				newOptGroupElement.appendChild(newOptionElement);
-				target.add(newOptGroupElement);
+			let optGroupElement = target.querySelector(`optgroup[label="${option.optgroup}"]`);
+			if (!optGroupElement) {
+				optGroupElement = document.createElement('optgroup');
+				if (optGroupElement instanceof HTMLOptGroupElement) {
+					optGroupElement.label = option.optgroup;
+					target.add(optGroupElement);
+				}
 			}
+			optGroupElement.appendChild(newOptionElement);
 		} else {
 			target.add(newOptionElement);
 		}
-		this.selectorElement.add(newOptionElement.cloneNode() as HTMLOptionElement);
+		return newOptionElement;
 	}
 
-	private async selectLeftScrolled() {
+	private selectLeftScrolled() {
 		if (!this.isIncomplete)
 			return;
 		const selectLeftScroll = this.selectLeftElement.scrollHeight - this.selectLeftElement.scrollTop;
 		if (selectLeftScroll <= this.selectLeftElement.offsetHeight) {
 			// triggers whenever the last <option>-element becomes visible inside its parent <select>
-			await this.remoteLookup();
+			this.remoteLookup();
 		}
 	}
 
-	private async remoteLookup() {
-		let query: string;
+	private remoteLookup() {
+		let query: URLSearchParams;
 		const searchString = this.searchLeftInput?.value;
 		if (searchString) {
 			const offset = this.selectLeftElement.querySelectorAll('option:not([hidden])').length;
-			query = `query=${searchString}&offset=${offset}`;
+			// query = `${this.buildFetchQuery(searchString)}&offset=${offset}`;
+			query = this.buildFetchQuery(offset, searchString);
 		} else {
-			query = `offset=${this.selectorElement.childElementCount}`;
+			// query = `${this.buildFetchQuery()}&offset=${this.selectorElement.childElementCount}`;
+			query = this.buildFetchQuery(this.selectorElement.childElementCount);
 		}
 		if (this.lastRemoteQuery === query)
 			return;
 		this.lastRemoteQuery = query;
-		await this.loadOptions(query, (options: Array<OptionData>) => options.forEach(option => {
-			this.addOptionToSelectElement(option, this.selectLeftElement);
+		this.loadOptions(query, (options: Array<OptionData>) => options.forEach(option => {
+			if (!(this.selectorElement.querySelector(`option[value="${option.id}"]`))) {
+				const optionElement = this.addOptionToSelectElement(option, this.selectLeftElement).cloneNode(false) as HTMLOptionElement;
+				this.selectorElement.add(optionElement);
+			}
+			this.setButtonsState();
 		}));
 	}
 
@@ -226,7 +219,7 @@ export class DualSelector extends IncompleteSelect {
 	private optionsSorted() {
 		this.selectorElement.querySelectorAll('option:checked').forEach(o => o.remove());
 		this.selectRightElement.querySelectorAll('option').forEach(optionElement => {
-			const clone = optionElement.cloneNode() as HTMLOptionElement;
+			const clone = optionElement.cloneNode(false) as HTMLOptionElement;
 			clone.selected = true;
 			this.selectorElement.add(clone);
 		});
@@ -255,21 +248,21 @@ export class DualSelector extends IncompleteSelect {
 		}
 	}
 
-	private async moveOptionRight(target: EventTarget | null) {
+	private moveOptionRight(target: EventTarget | null) {
 		if (target instanceof HTMLOptionElement) {
 			this.moveOptionToSelectElement(target, this.selectRightElement);
-			await this.selectLeftScrolled();
+			this.selectLeftScrolled();
 			this.hideOptionGroups(this.selectLeftElement);
 			this.optionsMoved();
 		}
 	}
 
-	private async moveAllOptionsRight() {
+	private moveAllOptionsRight() {
 		this.selectLeftElement.querySelectorAll('option:not([hidden])').forEach(option => {
 			this.moveOptionToSelectElement(option as HTMLOptionElement, this.selectRightElement);
 		});
 		this.clearSearchFields();
-		await this.selectLeftScrolled();
+		this.selectLeftScrolled();
 		this.optionsMoved();
 	}
 
@@ -311,14 +304,14 @@ export class DualSelector extends IncompleteSelect {
 		});
 	}
 
-	private async leftLookup() {
+	private leftLookup() {
 		const query = this.searchLeftInput?.value ?? '';
 		let numFoundOptions = this.lookup(this.selectLeftElement, query);
 		// first we look up for matching options ...
 		if (this.isIncomplete && numFoundOptions < this.selectLeftElement.size) {
 			// if we find less options than the <select> element can depict,
 			// query for additional matching options from the server.
-			await this.remoteLookup();
+			this.remoteLookup();
 			numFoundOptions = this.lookup(this.selectLeftElement, query);
 		}
 		this.setButtonsState();
@@ -435,9 +428,21 @@ export class DualSelector extends IncompleteSelect {
 		this.setHistoryCursor(0);
 	}
 
-	public getValue() : string[] {
-		const optionElements = Array.from(this.selectorElement.querySelectorAll('option:checked'));
-		return (optionElements as Array<HTMLOptionElement>).map(o => o.value);
+	protected reloadOptions() {
+		this.selectorElement.replaceChildren();
+		this.selectLeftElement.replaceChildren();
+		this.selectRightElement.replaceChildren();
+		this.clearSearchFields();
+		this.fieldGroup.classList.remove('dj-dirty', 'dj-touched', 'dj-validated');
+		this.fieldGroup.classList.add('dj-untouched', 'dj-pristine');
+		this.containerElement?.classList.remove('invalid');
+		const errorPlaceholder = this.fieldGroup.querySelector('.dj-errorlist > .dj-placeholder');
+		if (errorPlaceholder) {
+			errorPlaceholder.innerHTML = '';
+		}
+		this.remoteLookup();
+		this.historicalValues.splice(1);
+		this.setHistoryCursor(0);
 	}
 }
 
@@ -448,9 +453,5 @@ export class DualSelectorElement extends HTMLSelectElement {
 
 	connectedCallback() {
 		this[DS] = new DualSelector(this, 'django-dual-selector');
-	}
-
-	public get value() : any {
-		return this[DS]?.getValue();
 	}
 }