+++ testapp/test_e2e_fieldset.py
@@ -19,7 +19,7 @@ urlpatterns = [
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['customer'])
-def test_submit_customer(page, mocker):
+def test_submit_customer(page, mocker, viewname):
     customer_collections = page.query_selector_all('django-formset > django-form-collection')
     assert len(customer_collections) == 2
     fieldset = customer_collections[0].query_selector('fieldset[hide-if]')
@@ -40,7 +40,7 @@ def test_submit_customer(page, mocker):
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['customer'])
-def test_submit_no_customer(page, mocker):
+def test_submit_no_customer(page, mocker, viewname):
     assert page.query_selector('django-formset > django-form-collection:first-of-type fieldset[hide-if]') is not None
     page.click('#id_register\\.no_customer')
     assert page.query_selector('django-formset > django-form-collection:first-of-type fieldset[hidden]') is not None