+++ testapp/views.py
@@ -27,7 +27,7 @@ from formset.views import (
     FileUploadMixin, IncompleteSelectResponseMixin, FormCollectionView, FormViewMixin, EditCollectionView)
 
 from testapp.forms.address import AddressForm
-from testapp.forms.advertisement import AdvertisementForm, AdvertisementModelForm, initial_html
+from testapp.forms.advertisement import AdvertisementForm, AdvertisementModelForm
 from testapp.forms.article import ArticleForm
 from testapp.forms.complete import CompleteForm
 from testapp.forms.contact import (
@@ -39,6 +39,7 @@ from testapp.forms.opinion import OpinionForm
 from testapp.forms.person import ButtonActionsForm, SimplePersonForm, sample_person_data, ModelPersonForm
 from testapp.forms.poll import ModelPollForm, PollCollection
 from testapp.forms.questionnaire import QuestionnaireForm
+from testapp.forms.state import StateForm, StatesForm
 from testapp.forms.user import UserCollection, UserListCollection
 from testapp.forms.upload import UploadForm
 from testapp.models import AdvertisementModel, PersonModel, PollModel
@@ -454,6 +455,12 @@ urlpatterns = [
     path('counties', DemoFormView.as_view(
         form_class=CountyForm,
     ), kwargs={'group': 'form', 'index': 15}, name='counties'),
+    path('state', DemoFormView.as_view(
+        form_class=StateForm,
+    ), kwargs={'group': 'form', 'index': 15}, name='state'),
+    path('states', DemoFormView.as_view(
+        form_class=StatesForm,
+    ), kwargs={'group': 'form', 'index': 15}, name='states'),
     path('person', DemoModelFormView.as_view(
         form_class=ModelPersonForm,
         model=PersonModel,